straightned folder structure
