+++ client/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route} from "react-router-dom";
 import './App.css';
 import Navbar from "./components/Navbar"
-import Home from "./pages/home";
+import Home from "../src/pages/Home";
 
 function App() {
   return (
