+++ client/src/components/Col/index-check.js
@@ -3,5 +3,4 @@ import React from "react";
 function Row(props) {
     return <div className={`row${props.fluid ? "-fluid" : ""}`} {...props} />
 }
-
-export default Row;
+export default Col;
