+++ client/src/components/Col/index.js
@@ -1,5 +1,9 @@
 import React from "react";
 
-function Row(props) {
-    return <div className={`row${props.fluid ? "-fluid" : ""}`} {...props} />
+function Col(props) {
+  const size = props.size.split(" ").map(size => "col-" + size).join(" ");
+
+  return <div className={size} {...props} />;
 }
+
+export default Col;
