+++ client/src/components/Row/index.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Row(props) {
+    return <div className={`row${props.fluid ? "-fluid" : ""}`} {...props} />
+}
+
+export default Row;
