--- src/components/Col/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-function Col(props) {
-  const size = props.size.split(" ").map(size => "col-" + size).join(" ");
-
-  return <div className={size} {...props} />;
-}
-
-export default Col;
